add more convenient API to derive a module descriptor from root class
In cases where a certain descriptor root class of a module is used to define the module it makes sense to offer a convenient API to directly derive the module descriptor from this root class instead of having to search the root class again within all the classes contained within the module. One example is if the root class is annotated with a certain annotation containing the module meta-information. In this case we need the root class again to derive the module descriptor.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>