add `modules()...onlyDependOnEachOtherThroughClassesThat(predicate)`
This is another common requirement when testing for modularization, that no matter the visibility of individual classes the API of each module is a dedicated subset of all (compile) accessible classes. The reason is that the Java visibility modifiers are typically not powerful enough to model such a module API, unless all classes are thrown together in a single package (where we can use package-visibility), which is not feasible for more complex modules.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>